Fix corner case in text editor extra area selection highlight range
In a multi-line selection if the cursor is at the start of the line
then highlight that line as well, to indicate that the line ending of
the previous line got selected, even if nothing of the current line is
selected. This better matches the number of lines that is actually
produces if the selection is copy pasted to somewhere else.