Deal with line height vs line spacing problem
The line height is meant to be high enough to contain text without
clipping it. The line spacing is the distance between two base lines.
The line spacing is defined as the line height plus the line leading.
Typically the leading should not be negative, which would make the
line spacing smaller than the line height resulting in overlapped or
clipped text.

Negative line leading doesn't seem to be much of a problem on Windows
but seems to be common on Linux. Therefore, don't use the line spacing
but use the line height for the vertical text layout.