Detect decoding errors during text document loading
Show decoding errors in a new infobar at the top of the document.
The infobar also contains a button to open a new dialog to select
a different codec to reopen the text document with.

Also move the ownership of documents from the document manager to
the corresponding editors to simplify ownership.