Restructure file opening logic to finish encoding handling
Add TextCodec class that also includes Unicode BOM handling to
avoid having to pass around a QTextCodec and a bool value.

Move encoding-change logic to DocumentManager to avoid code
duplication between MainWindow and TextEditorWidget classes.

Move file opening logic from Document subclasses to the
DocumentManager class as it is always the same and doesn't
depend on the document type.